Initial tidy up and move to conduits
I've tidied up the code to conform to my general style - 80 columns, no trailing
whitespace, etc. I know this is really unprofessional to do to someone elses
work, but Michael has over ownership to me, and as I plan to maintain this for a
while, I think it's justified. Best to get this in earlier, rather than break
blame later!

Of more notable changes, I've cleanup the .cabal specification, and removed the
enumerator code in favour of conduits. The FromXML stuff has been removed to
live in Text.XML.Cursor.FromXML, with the possibility that this can be shipped
in the future.